Fix typos

+++ firmware/i2c_operations.h
@@ -9,5 +9,5 @@ typedef unsigned char UINT8;
 void i2c_write8(UINT8 addr, UINT8 reg, UINT8 value);
 void i2c_write16(UINT8 address, UINT8 reg, UINT16 value);
 UINT8 i2c_read8(UINT8 addr, UINT8 reg);
-UINT8 i2c_read16(UINT8 addr, UINT8 reg);
+UINT16 i2c_read16(UINT8 addr, UINT8 reg);
 UINT8 i2c_readBuffer(UINT8 addr, UINT8 reg, UINT8 len, UINT8 *buffer);
